Add BaseBroadcastService pr comment adjustments
